task2

+++ 0x0A-argc_argv/4-add.c
@@ -14,7 +14,7 @@ int main(int argc, char *argv[])
 
 	int i;
 	int sum = 0;
-	long temp;
+	int temp;
 	char *endptr;
 
 	if (argc == 1)
